+++ src/core/atom-class.ts
@@ -90,6 +90,7 @@ export type AtomType =
   // by environments such as `matrix`, `cases`, etc...
   | 'box' // A border drawn around an expression and change its background color
   | 'chem' // A chemical formula (mhchem)
+  | 'choice' // A \\mathchoice command
   | 'composition' // IME composition area
   | 'delim'
   | 'enclose'